Refactor Daily and Weekly Price Calculation and update AGENTS.md guidelines (#134)
* manual code formatting

* Refactor Yahoo Finance JSON parser to YahooResponseParser and decouple fetching, hardcode fallback date in DailyPriceRepository, and document learnings

* Update backend/AGENTS.md with AI learnings and coding guidelines

* Rename IndicatorProperties to IndicatorConfig and update all references

* Refactor WeeklyPriceCalculator to loop through all active tickers without pagination and merge WeeklyTickerProcessor logic

* Remove default fallback date and skip weekly price calculation when no daily prices exist

* Use findLatestPriceDatesForActiveTickers instead of findTopByTickerOrderByPriceDateAsc in WeeklyPriceCalculator

* Remove TickerRepository dependency and combine processTicker methods in WeeklyPriceCalculator

* Refactor high and low weekly price aggregation using Stream max and min

* Update AGENTS.md and backend/AGENTS.md with new AI learnings on transactional self-proxy fallback, N+1 bulk query elimination, and stream aggregations

* Renamed a function

+++ backend/src/main/java/com/alphaflow/api/dtos/OhlcvDTO.java
@@ -12,4 +12,4 @@ public record OhlcvDTO(
     @JsonProperty("high") BigDecimal priceHigh,
     @JsonProperty("low") BigDecimal priceLow,
     @JsonProperty("close") BigDecimal priceClose,
-    @JsonProperty("vol") Long volume) {}
+    @JsonProperty("vol") BigDecimal volume) {}
